# Rewrite vector_addition function to a 4x faster recursive version