Hi,

This vector addition function is 4x faster.
I have pushed the benchmark code to my fork in the same `bench_new_hamming` branch I used for the hamming_distance benchmark.
You can run it with `mix run bench/bench_vector_addition.exs` in that branch.

```
Benchmark suite executing with the following configuration:
warmup: 2 s
time: 5 s
memory time: 0 ns
parallel: 1
inputs: none specified
Estimated total run time: 14 s

Benchmarking vector_addition...
Benchmarking vector_addition2...

Name                       ips        average  deviation         median         99th %
vector_addition2        265.37        3.77 ms     ±5.29%        3.71 ms        4.53 ms
vector_addition          68.36       14.63 ms    ±10.60%       14.15 ms       18.67 ms

Comparison: 
vector_addition2        265.37
vector_addition          68.36 - 3.88x slower +10.86 ms
```

:wave: